test_requirements.txt
factory_boy==2.8.1
freezegun==0.3.8
Faker==0.7.10
pytest==3.0.7
testing.postgresql==1.3.0
